+++ montessori-os/src/components/VersionBadge.jsx
@@ -48,7 +48,7 @@ const VersionBadge = ({ userRole, showInProfile = false }) => {
             fontSize: '1.1rem'
           }}
         >
-          v11.5.0
+          v11.6.0
         </Typography>
       </Box>
     );
@@ -88,7 +88,7 @@ const VersionBadge = ({ userRole, showInProfile = false }) => {
           fontSize: 'inherit'
         }}
       >
-        v11.5.0
+        v11.6.0
       </Typography>
     </Box>
   );